fix(menu): fix an issue when the menu items are duplicating
